Hide example screencast in collapsible block
The embedded example.svg creates some INCREDIBLY bad lag in Chrome;
enough that simply having the README open causes the tab to be nearly
unusable.

It doesn't even load in Safari. Firefox seems fine for some reason.

The lag in Chrome only exists when the SVG is on-screen, not when it
loads. This is obviously a stopgap; I would consider moving that example to
a screencast host and embedding that instead.