fix(bash): allow proper completion of -n and -e flags
Fix issue where -n and -e flags were being silently stripped in
_fzf_bash_completion_flatten_subshells. The consequence of this was
that completion would not work for "-e" or "-n" flags. For example:

kubectl get pod -n <TAB> -> would try to complete on "pod"
kubectl get pod -n=<TAB> -> would correctly complete on "-n"

The cause was that bash's echo command will print an empty line if you
give it the literal strings '-e' or '-n' as sole input. The fix is to
use printf, which should reliably print the input it receives.

I've changed most of the echo calls in bash-completion.sh to use
printf, to reduce risk that they hit the same issue.